changed how markers positions are tracked and displayed
The robot's frame of reference (odom) is different from the map (map). The robot thinks it's starting at (0,0) when it's starting at map position (-1,3), so the two frames are actually translated and rotated from each other. This change tracks the markers position in odom frame so the robot knows where they are properly, but tracks their visual appearance in the map frame so that they display properly. It's a kludgy fix for now and should be improved.